refactor: split workspace log/artifact/provenance schemas into sibling module
The opt-in auto_merge feature pushed src/awf/api/schemas.py to 1528 lines,
over the 1500-line first-party maintainability guard. Move the self-contained
leaf read models (workspace log/artifact/validation-provenance responses,
which depend only on stdlib + schemas_operations aliases, never on any class
defined in schemas) into schemas_workspace_io.py and re-export them, matching
the existing schemas_operations/schemas_responses pattern. No behavior change;
openapi.json is unaffected.
